move components directory out of pages
