+++ web/pages/index.tsx
@@ -1,8 +1,8 @@
 import type { NextPage } from "next";
 import { useState } from "react";
 
-import { GuessForm } from "./components/guessForm";
-import { GuessTable } from "./components/guessTable";
+import { GuessForm } from "../components/guessForm";
+import { GuessTable } from "../components/guessTable";
 
 export interface guessResult {
   guess: string;
